+++ apps/angular-example/src/app/test_components/a.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'a-component',
-  template: '<div>TEST A</div>',
-  styles: [''],
-})
-export class ATestComponent {}