+++ apps/angular-example/src/app/test_components/b.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'b-component',
-  template: '<div>TEST B</div>',
-  styles: [''],
-})
-export class BTestComponent {}