+++ apps/angular-example/src/environments/environment.ts
@@ -1,5 +1,0 @@
-export const environment = {
-  production: true,
-  ABBY_PROJECT_ID:
-    process.env['ABBY_PROJECT_ID'] ?? 'clftg3tzd0004l7085yktpsov',
-};