+++ packages/angular/src/lib/abby.service.ts
@@ -19,7 +19,6 @@ import {
   switchMap,
   tap,
 } from "rxjs";
-import { F } from "ts-toolbelt";
 import { AbbyLoggerService } from "./abby-logger.service";
 import {
   FlagStorageService,
@@ -44,9 +43,8 @@ type LocalData<
 };
 
 export type InferFlagNames<C extends AbbyConfig> = InferFlags<C>[number];
-export type InferTestNames<C extends AbbyConfig> = InferTests<C> extends Record<infer T, any>
-  ? T
-  : never;
+export type InferTestNames<C extends AbbyConfig> =
+  InferTests<C> extends Record<infer T, any> ? T : never;
 export type InferTests<C extends AbbyConfig> = NonNullable<C["tests"]>;
 export type InferFlags<C extends AbbyConfig> = NonNullable<C["flags"]>;
 export type InferRemoteConfig<C extends AbbyConfig> = NonNullable<C["remoteConfig"]>;
@@ -56,14 +54,14 @@ type PossibleFlagName<FlagName extends string> = FlagName | `!${FlagName}`;
 
 @Injectable({ providedIn: "root" })
 export class AbbyService<
-  FlagName extends string = string,
-  TestName extends string = string,
-  Tests extends Record<TestName, ABConfig> = Record<TestName, ABConfig>,
-  RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString> = Record<
+  const FlagName extends string = string,
+  const TestName extends string = string,
+  const Tests extends Record<TestName, ABConfig> = Record<TestName, ABConfig>,
+  const RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString> = Record<
     string,
     RemoteConfigValueString
   >,
-  RemoteConfigName extends Extract<keyof RemoteConfig, string> = Extract<
+  const RemoteConfigName extends Extract<keyof RemoteConfig, string> = Extract<
     keyof RemoteConfig,
     string
   >,
@@ -72,7 +70,7 @@ export class AbbyService<
 
   private selectedVariants: { [key: string]: string } = {};
 
-  private config: F.Narrow<AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>>;
+  private config: AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>;
 
   private projectData$?: Observable<LocalData<FlagName, TestName>>;
 
@@ -80,10 +78,10 @@ export class AbbyService<
 
   constructor(
     @Inject(AbbyService)
-    config: F.Narrow<AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>>,
+    config: AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>,
     private abbyLogger: AbbyLoggerService
   ) {
-    this.abby = new Abby<FlagName, TestName, Tests, RemoteConfig, RemoteConfigName>(
+    this.abby = new Abby(
       config,
       {
         get: (key: string) => {
@@ -127,28 +125,13 @@ export class AbbyService<
     return this.resolveData().pipe(map(() => void 0));
   }
 
-  public getVariant<T extends keyof Tests>(testName: T): Observable<string>;
-  public getVariant<T extends keyof Tests, S>(
-    testName: T,
-    lookupObject: { [key in ExtractVariants<T, Tests>]: S } | undefined
-  ): Observable<S>;
-  public getVariant<T extends keyof Tests, S>(
-    testName: T,
-    lookupObject?: { [key in ExtractVariants<T, Tests>]: S } | undefined
-  ): Observable<string | S> {
+  public getVariant<T extends keyof Tests>(testName: T): Observable<string> {
     this.abbyLogger.log(`getVariant(${testName as string})`);
 
     return this.resolveData().pipe(
       map((data) => this.abby.getTestVariant(testName)),
       tap((variant) => (this.selectedVariants[testName as string] = variant)),
       tap((variant) => this.abbyLogger.log(`getVariant(${testName as string}) =>`, variant)),
-      map((variant) => {
-        if (lookupObject === undefined) {
-          return variant;
-        }
-
-        return lookupObject[variant];
-      }),
       shareReplay(1)
     );
   }