+++ packages/next/src/index.tsx
@@ -6,7 +6,6 @@ import {
   ABTestReturnValue,
 } from "@tryabby/react";
 import { AbbyDataResponse, getABStorageKey, RemoteConfigValueString } from "@tryabby/core";
-import type { F } from "ts-toolbelt";
 import { ABBY_DATA_KEY, withAbby } from "./withAbby";
 import { HttpService } from "@tryabby/core";
 import type { NextMiddleware, NextRequest, NextResponse } from "next/server";
@@ -18,19 +17,12 @@ export { defineConfig } from "@tryabby/core";
 
 // TODO: figure out how to prevent re-typing of the same types
 export function createAbby<
-  FlagName extends string,
-  TestName extends string,
-  Tests extends Record<TestName, ABConfig>,
-  RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
-  RemoteConfigName extends Extract<keyof RemoteConfig, string>,
-  ConfigType extends AbbyConfig<
-    FlagName,
-    Tests,
-    string[],
-    RemoteConfigName,
-    RemoteConfig
-  > = AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>,
->(config: F.Narrow<AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>>) {
+  const FlagName extends string,
+  const TestName extends string,
+  const Tests extends Record<string, ABConfig>,
+  const RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
+  const RemoteConfigName extends Extract<keyof RemoteConfig, string>,
+>(config: AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>) {
   const {
     AbbyProvider,
     useAbby,
@@ -43,7 +35,7 @@ export function createAbby<
     withDevtools,
     useFeatureFlags,
     useRemoteConfigVariables,
-  } = baseCreateAbby<FlagName, TestName, Tests, RemoteConfig, RemoteConfigName, ConfigType>(config);
+  } = baseCreateAbby<FlagName, TestName, Tests, RemoteConfig, RemoteConfigName>(config);
 
   const abbyApiHandler =
     <HandlerType extends NextApiHandler | NextMiddleware>(handler: HandlerType) =>
@@ -61,21 +53,7 @@ export function createAbby<
 
   return {
     AbbyProvider,
-    // we need to retype the useAbby function here
-    // because re-using the types from the react package causes the ts-toolbelt package to behave weirdly
-    // and therefore not working as expected
-    useAbby: useAbby as <
-      K extends keyof Tests,
-      TestVariant extends Tests[K]["variants"][number],
-      LookupValue,
-      Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
-    >(
-      name: K,
-      lookupObject?: F.Narrow<Lookup>
-    ) => {
-      variant: ABTestReturnValue<Lookup, TestVariant>;
-      onAct: () => void;
-    },
+    useAbby,
     useFeatureFlag,
     withAbby: withAbby(config as AbbyConfig<FlagName, Tests>, __abby__),
     getFeatureFlagValue,
@@ -100,7 +78,7 @@ export function createAbby<
       T extends keyof Tests,
       TestVariant extends Tests[T]["variants"][number],
       LookupValue,
-      Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
+      const Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
       RequestType extends NextRequest | NextApiRequest | undefined = undefined,
       ResponseType extends NextResponse | NextApiResponse = RequestType extends NextRequest
         ? NextResponse
@@ -110,7 +88,7 @@ export function createAbby<
     >(
       name: T,
       req?: RequestType,
-      lookupObject?: F.Narrow<Lookup>
+      lookupObject?: Lookup
     ): [
       Lookup extends undefined
         ? TestVariant
@@ -129,7 +107,7 @@ export function createAbby<
         const storedVariant = typeof storedValue === "string" ? storedValue : storedValue.value;
 
         return [
-          lookupObject ? lookupObject[storedVariant as keyof typeof lookupObject] : storedVariant,
+          lookupObject ? lookupObject[storedVariant as TestVariant] : storedVariant,
           () => {},
         ];
       }
@@ -158,7 +136,7 @@ export function createAbby<
       };
 
       return [
-        lookupObject ? (lookupObject[newValue as keyof typeof lookupObject] as any) : newValue,
+        lookupObject ? (lookupObject[newValue as TestVariant] as any) : newValue,
         setCookieFunc,
       ];
     },