+++ packages/remix/src/index.tsx
@@ -6,7 +6,6 @@ import {
   ABTestReturnValue,
 } from "@tryabby/react";
 import { AbbyDataResponse, HttpService, RemoteConfigValueString } from "@tryabby/core";
-import type { F } from "ts-toolbelt";
 import { useMatches } from "@remix-run/react";
 
 const ABBY_DATA_KEY = "__ABBY_PROJECT_DATA__";
@@ -19,19 +18,12 @@ import { PromiseCache } from "./cache";
 export { defineConfig } from "@tryabby/core";
 
 export function createAbby<
-  FlagName extends string,
-  TestName extends string,
-  Tests extends Record<TestName, ABConfig>,
-  RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
-  RemoteConfigName extends Extract<keyof RemoteConfig, string>,
-  ConfigType extends AbbyConfig<
-    FlagName,
-    Tests,
-    string[],
-    RemoteConfigName,
-    RemoteConfig
-  > = AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>,
->(config: F.Narrow<AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>>) {
+  const FlagName extends string,
+  const TestName extends string,
+  const Tests extends Record<TestName, ABConfig>,
+  const RemoteConfig extends Record<RemoteConfigName, RemoteConfigValueString>,
+  const RemoteConfigName extends Extract<keyof RemoteConfig, string>,
+>(config: AbbyConfig<FlagName, Tests, string[], RemoteConfigName, RemoteConfig>) {
   const {
     AbbyProvider,
     useAbby,
@@ -45,7 +37,7 @@ export function createAbby<
     withDevtools,
     useFeatureFlags,
     useRemoteConfigVariables,
-  } = baseCreateAbby<FlagName, TestName, Tests, RemoteConfig, RemoteConfigName, ConfigType>(config);
+  } = baseCreateAbby<FlagName, TestName, Tests, RemoteConfig, RemoteConfigName>(config);
 
   const AbbyRemixProvider = ({ children }: PropsWithChildren) => {
     const matches = useMatches();
@@ -104,21 +96,7 @@ export function createAbby<
 
   return {
     AbbyProvider: AbbyRemixProvider,
-    // we need to retype the useAbby function here
-    // because re-using the types from the react package causes the ts-toolbelt package to behave weirdly
-    // and therefore not working as expected
-    useAbby: useAbby as <
-      K extends keyof Tests,
-      TestVariant extends Tests[K]["variants"][number],
-      LookupValue,
-      Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
-    >(
-      name: K,
-      lookupObject?: F.Narrow<Lookup>
-    ) => {
-      variant: ABTestReturnValue<Lookup, TestVariant>;
-      onAct: () => void;
-    },
+    useAbby,
     useFeatureFlag,
     getFeatureFlagValue,
     useRemoteConfig,