+++ packages/node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tryabby/node
 
+## 6.0.0
+
+### Major Changes
+
+- use typescript 5
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@6.0.0
+
 ## 5.1.8
 
 ### Patch Changes